+++ stl/lib/stl.dart
@@ -3,5 +3,4 @@
 /// More dartdocs go here.
 library;
 
-export 'src/stl_base.dart';
-export 'src/vector/vector.dart';
+export 'src/collections/vector.dart';