+++ stl/CHANGELOG.md
@@ -49,3 +49,8 @@
 - Added `removeRange()` method for removing a range of elements.
 - Added `removeWhere()` method for removing elements that satisfy a condition.
 - Added `retainWhere()` method for retaining elements that satisfy a condition.
+
+## 0.1.4
+- Added `stl.collections` for more collection types.
+- Moved `Vector` to `stl.collections`.
+- Removed `stl_base.dart`.